requirements.txt
# Core AI/LLM
langchain==0.3.13
langchain-google-genai==2.0.8
langgraph==0.2.62
langchain-core==0.3.28
langchain-community==0.3.13

# Git operations
gitpython==3.1.43
pygithub==2.5.0
python-gitlab==4.14.0

# Code analysis
pylint==3.3.4
flake8==7.1.1
black==24.10.0
mypy==1.14.0
bandit==1.8.0
radon==6.0.1
astroid==3.3.8
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1

# Security scanning
safety==3.2.11
semgrep==1.103.0
pip-audit==2.8.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
coverage==7.6.9

# Performance profiling
memory-profiler==0.61.0
py-spy==0.4.0

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
rich==13.9.4
click==8.1.8
tenacity==9.0.0
pydantic==2.10.5

# Reporting
jinja2==3.1.5
markdown==3.7
reportlab==4.2.5
